orchestrator/requirements.txt
# Orchestrator runtime dependencies
fastapi>=0.111
uvicorn>=0.30
httpx>=0.27
pydantic>=2.6
email-validator>=2.1.0
opensearch-py>=2.5
tenacity>=8.5
pyyaml>=6.0.2

# ingest/file_meta deps
Pillow>=10.3
trafilatura>=1.8
pypdf>=4.2

# OSINT tools (CLI usage via subprocess, αλλά τα κρατάμε εδώ για ευκολία)
maigret==0.5.0
holehe
